Add ordered version of the operator
