+++ stream/src/main/scala/org/apache/pekko/stream/javadsl/Source.scala
@@ -2496,13 +2496,11 @@ final class Source[Out, Mat](delegate: scaladsl.Source[Out, Mat]) extends Graph[
   /**
    * @since 1.1.0
    */
-  def mapAsyncPartitioned[T, P](parallelism: Int,
+  def mapAsyncPartitionedUnordered[T, P](parallelism: Int,
       bufferSize: Int,
       extractPartition: function.Function[Out, P],
-      f: function.Function2[Out, P, CompletionStage[T]]): javadsl.Source[T, Mat] = {
-    MapAsyncPartition.mapSourceAsyncPartition(delegate, parallelism, bufferSize)(extractPartition(_))(f(_, _).asScala)
-      .asJava
-  }
+      f: function.Function2[Out, P, CompletionStage[T]]): javadsl.Source[T, Mat] =
+    MapAsyncPartitioned.mapSourceUnordered(delegate, parallelism, bufferSize)(extractPartition(_))(f(_, _).asScala).asJava
 
   /**
    * Transform this stream by applying the given function to each of the elements
