+++ stream/src/main/scala/org/apache/pekko/stream/scaladsl/FlowWithContext.scala
@@ -94,10 +94,20 @@ final class FlowWithContext[-In, -CtxIn, +Out, +CtxOut, +Mat](delegate: Flow[(In
    * @since 1.1.0
    */
   def mapAsyncPartitioned[T, P](parallelism: Int,
-      bufferSize: Int = MapAsyncPartition.DefaultBufferSize)(
+      bufferSize: Int = MapAsyncPartitioned.DefaultBufferSize)(
       extractPartition: Out => P)(
       f: (Out, P) => Future[T]): FlowWithContext[In, CtxIn, T, CtxOut, Mat] = {
-    MapAsyncPartition.mapFlowWithContextAsyncPartition(this, parallelism, bufferSize)(extractPartition)(f)
+    MapAsyncPartitioned.mapFlowWithContextOrdered(this, parallelism, bufferSize)(extractPartition)(f)
+  }
+
+  /**
+   * @since 1.1.0
+   */
+  def mapAsyncPartitionedUnordered[T, P](parallelism: Int,
+      bufferSize: Int = MapAsyncPartitioned.DefaultBufferSize)(
+      extractPartition: Out => P)(
+      f: (Out, P) => Future[T]): FlowWithContext[In, CtxIn, T, CtxOut, Mat] = {
+    MapAsyncPartitioned.mapFlowWithContextUnordered(this, parallelism, bufferSize)(extractPartition)(f)
   }
 
   def asFlow: Flow[(In, CtxIn), (Out, CtxOut), Mat] = delegate
